[Swift 4] Update Shuffle to Swift 4
